+++ patzilla-ui/navigator/style/index.js
@@ -1,1 +1,2 @@
+require('./generic.css');
 require('./app.css');