+++ patzilla-ui/navigator/util/linkmaker.js
@@ -191,7 +191,7 @@
             // http://ccd.fiveipoffices.org/CCD-2.0/html/viewCcd.html?num=DE20132003344U&type=application&format=epodoc
             // http://ccd.fiveipoffices.org/CCD-2.0/html/viewCcd.html?num=US201113881490&type=application&format=epodoc
             var document_id = this.document.get_application_reference('epodoc');
-            var url_tpl = _.template('http://ccd.fiveipoffices.org/CCD-2.1.6/html/viewCcd.html?num=<%= fullnumber %>&type=application&format=epodoc');
+            var url_tpl = _.template('http://ccd.fiveipoffices.org/CCD-2.1.8/html/viewCcd.html?num=<%= fullnumber %>&type=application&format=epodoc');
             var url = url_tpl(document_id);
             return url;
         },