+++ CHANGES.rst
@@ -6,6 +6,7 @@ IP Navigator CHANGES
 Development
 ===========
 - [ui] Bump wording
+- [ui] Fix conditional menu item display
 
 
 2018-11-20 0.159.0