+++ CHANGES.rst
@@ -8,6 +8,7 @@ Development
 - [ui] Bump wording
 - [ui] Fix conditional menu item display
 - [ui] Improve design and positioning of menu opener
+- [ui] Refactor main application CSS
 
 
 2018-11-20 0.159.0