+++ CHANGES.rst
@@ -9,6 +9,7 @@ Development
 - [ui] Fix conditional menu item display
 - [ui] Improve design and positioning of menu opener
 - [ui] Refactor main application CSS
+- [ui,mw] Refactor vendor-specific feature flags
 
 
 2018-11-20 0.159.0